Create README.md
I am a Danish based SEO professional, and i am currently learning coding for my future career. 

My goal is to become a developer and continue to expand my knowledge of web applications, digitalization, AI integration and keep up with the development as an employee with coworkers to spar and develop with.